fix(normalize): move SSoT out of deprecated core
