+++ internal/control/http/v3/handlers_intents.go
@@ -15,11 +15,11 @@ import (
 	"github.com/google/uuid"
 
 	"github.com/ManuGH/xg2g/internal/control/admission"
-	"github.com/ManuGH/xg2g/internal/core/normalize"
 	"github.com/ManuGH/xg2g/internal/domain/session/lifecycle"
 	"github.com/ManuGH/xg2g/internal/domain/session/model"
 	"github.com/ManuGH/xg2g/internal/log"
 	"github.com/ManuGH/xg2g/internal/metrics"
+	"github.com/ManuGH/xg2g/internal/normalize"
 	v3api "github.com/ManuGH/xg2g/internal/pipeline/api"
 	"github.com/ManuGH/xg2g/internal/pipeline/hardware"
 	"github.com/ManuGH/xg2g/internal/pipeline/profiles"
